fix(chat): User not being able to delete a message using keyboard navigation only
